start refactoring core/ and cogs/
