+++ bot.py
@@ -19,27 +19,12 @@ bot = commands.Bot(command_prefix=f'{PREFIX} ', case_insensitive=True,
                    description='Seamless GitHub-Discord integration.',
                    fetch_offline_members=False)
 
-dir_paths: list = ['./cogs', './handle', './ext', './core']
-exceptions: list = ["explicit_checks.py", "decorators.py", "manager.py", "api.py"]
-staging_exceptions: list = ['topgg.py'] if not PRODUCTION else []
-botlist_folders: list = []
-
 logging.basicConfig(level=logging.INFO, format='%(levelname)s:%(name)s: %(message)s')
 logging.getLogger('asyncio').setLevel(logging.WARNING)
 logging.getLogger('discord.gateway').setLevel(logging.WARNING)
 logger = logging.getLogger('main')
 
-bot.logger = logger
-
-
-# Before a command is ran
-async def before_invoke(ctx: commands.Context):
-    await ctx.channel.trigger_typing()
-
-
-bot.before_invoke(before_invoke)
-
-extensions = [
+extensions: list = [
     'cogs.base.user',
     'cogs.base.org',
     'cogs.base.repo',
@@ -51,24 +36,16 @@ extensions = [
     'cogs.help',
     'cogs.config',
     'cogs.debug',
-    'cogs.bot_info'
+    'cogs.bot_info',
+    'handle.errors',
+    'handle.events'
 ]
 
-for extension in extensions:
+botlist_extensions: list = [f'cogs.botlists.{file[:-3]}' for file in os.listdir('cogs/botlists')] if PRODUCTION else []
+
+for extension in (extensions + botlist_extensions):
     logger.info(f'Loading {extension}...')
     bot.load_extension(extension)
-
-if PRODUCTION:  # Load botlist extensions if we are in a production environment
-    for folder in os.listdir('./core/botlists'):
-        if os.path.isdir(f'./core/botlists/{folder}'):
-            logger.info(f'queueing {folder} to be loaded')
-            botlist_folders.append((f'./core/botlists/{folder}', f"core.botlists.{folder}"))
-
-    for folder in botlist_folders:
-        for file in os.listdir(folder[0]):
-            if file.endswith('.py'):
-                logger.info(f'loading extension: {folder[1]}.{file[:-3]}')
-                bot.load_extension(f"{folder[1]}.{file[:-3]}")
 
 
 @bot.check
@@ -79,8 +56,15 @@ async def global_check(ctx: commands.Context) -> bool:
     return True
 
 
+async def before_invoke(ctx: commands.Context) -> None:
+    await ctx.channel.trigger_typing()
+
+
+bot.before_invoke(before_invoke)
+
+
 @bot.event
-async def on_ready():
+async def on_ready() -> None:
     logger.info(f'The bot is ready.')
     logger.info(f'discord.py version: {discord.__version__}\n')
